+++ guerilla/players.py
@@ -129,7 +129,6 @@ class Guerilla(Player):
 class Human(Player):
     def __init__(self, name, colour=None):
         super(Human, self).__init__(name)
-        self.gui = None
 
     def __enter__(self):
         return self
@@ -137,12 +136,6 @@ class Human(Player):
     def __exit__(self, e_type, value, traceback):
         if e_type is not None:
             print e_type, value, traceback
-
-    def get_move_from_gui(self, board):
-        move = self.gui.get_player_input(board)
-        if move == "undo":
-            return move
-        return chess.Move.from_uci(move)
 
     def get_move_from_tml(self, board):
         move = None
@@ -167,50 +160,14 @@ class Human(Player):
 
         return move
 
-    def get_move(self, board, gui=True):
+    def get_move(self, board):
         """
         Fetches a move from the command line.
         Input:
             Board [Chess.board]
         """
+        return self.get_move_from_tml(board)
 
-        if gui:
-            return self.get_move_from_gui(board)
-        else:
-            return self.get_move_from_tml(board)
-
-try:
-    import other_engines.sunfish.sunfish as sunfish
-    import other_engines.sunfish.tools as sun_tools
-except ImportError:
-    pass
-else:
-    class Sunfish(Player):
-        def __init__(self, name, colour=None, time_limit=2):
-            """
-            Initializes the Sunfish class.
-            Input:
-                time_limit [Integer]
-                    The time limit for each move search in SECONDS.
-            """
-            super(Sunfish, self).__init__(name)
-            self.search = sunfish.Searcher()
-            self.time_limit = time_limit
-
-        def __enter__(self):
-            return
-
-        def __exit__(self, e_type, value, traceback):
-            if e_type is not None:
-                print e_type, value, traceback
-
-        def get_move(self, board):
-            # Convert to Sunfish position
-            sun_pos = sun_tools.parseFEN(board.fen())
-
-            # Returns move
-            move, _ = self.search.search(sun_pos, self.time_limit)
-            return board.parse_san(sun_tools.renderSAN(sun_pos, move))
 
 class Stockfish(Player):
     def __init__(self, name, colour=None, time_limit=2):